cambio en migraciones y cambio en migracion notificacion

+++ database/migrations/2022_05_24_172047_create_notificaciones.php
@@ -18,7 +18,7 @@ class CreateNotificaciones extends Migration
             $table->string('email');
             $table->string('mensaje');
             $table->integer('leido');
-            $table->integer('semana');
+            $table->unsignedBigInteger('semana');
             $table->timestamps();
 
             $table->foreign('email')->references('email')->on('users');
